fix(validation): extract isStrictDateString, extend round-trip to datetimes
Cubic round 2: 2024-02-31T00:00:00Z was accepted (Date.parse rolls over).
- Extracted isStrictDateString to lib/validation/common.ts (shared)
- Datetime branch now extracts date components and applies the same
  round-trip check as date-only: rejects 2024-02-31T00:00:00Z and
  2023-02-29T12:00:00+01:00
- Tests import from shared location (no duplication)
- 2 new RED→GREEN tests for datetime round-trip
