+++ __tests__/api/invoices.issuedAt.test.ts
@@ -3,28 +3,11 @@
  *
  * Cubic cases: 2024-02-31 rolls to March, 2023-02-29 is not a leap year,
  * timezone-less datetimes are environment-dependent.
+ * Round 2: datetime round-trip (2024-02-31T00:00:00Z, 2023-02-29T12:00:00+01:00).
  */
 
 import { z } from 'zod';
-
-// Extract the validation logic to test it directly
-function isStrictDateString(v: string): boolean {
-  const dateOnlyMatch = v.match(/^(\d{4})-(\d{2})-(\d{2})$/);
-  if (dateOnlyMatch) {
-    const [, yearStr, monthStr, dayStr] = dateOnlyMatch;
-    const year = Number(yearStr);
-    const month = Number(monthStr);
-    const day = Number(dayStr);
-    const d = new Date(year, month - 1, day);
-    return d.getFullYear() === year && d.getMonth() === month - 1 && d.getDate() === day;
-  }
-  const datetimeMatch = v.match(/^(\d{4})-(\d{2})-(\d{2})T.+$/);
-  if (datetimeMatch) {
-    if (!/[Zz]$|[+-]\d{2}:\d{2}$/.test(v)) return false;
-    return !isNaN(Date.parse(v));
-  }
-  return false;
-}
+import { isStrictDateString } from '@/lib/validation/common';
 
 const strictDateSchema = z.string().refine(isStrictDateString, {
   message: 'Date invalide',
@@ -80,6 +63,16 @@ describe('issuedAt strict date validation', () => {
 
     it('rejects timezone-less datetime with milliseconds', () => {
       expect(strictDateSchema.safeParse('2024-01-01T12:00:00.000').success).toBe(false);
+    });
+  });
+
+  describe('datetime round-trip (rejects rolled-over datetimes)', () => {
+    it('rejects 2024-02-31T00:00:00Z (Feb 31 rolls to March)', () => {
+      expect(strictDateSchema.safeParse('2024-02-31T00:00:00Z').success).toBe(false);
+    });
+
+    it('rejects 2023-02-29T12:00:00+01:00 (2023 is not a leap year)', () => {
+      expect(strictDateSchema.safeParse('2023-02-29T12:00:00+01:00').success).toBe(false);
     });
   });
