+++ lib/validation/common.ts
@@ -75,3 +75,38 @@ export const optionalString = z.string().trim().optional().or(z.literal('').tran
  * Required non-empty string
  */
 export const nonEmptyString = z.string().trim().min(1, 'This field is required');
+
+/**
+ * Strict date string validator:
+ * - YYYY-MM-DD: round-trip check (parsed year/month/day must match input components)
+ *   to reject rolled-over dates like 2024-02-31, 2023-02-29
+ * - Datetime: must include timezone offset or Z (rejects timezone-less like 2024-01-01T12:00:00),
+ *   and applies the same round-trip check on the date portion to reject rolled-over datetimes
+ *   like 2024-02-31T10:30:00Z
+ */
+export function isStrictDateString(v: string): boolean {
+  const dateOnlyMatch = v.match(/^(\d{4})-(\d{2})-(\d{2})$/);
+  if (dateOnlyMatch) {
+    const [, yearStr, monthStr, dayStr] = dateOnlyMatch;
+    return isValidDateComponents(Number(yearStr), Number(monthStr), Number(dayStr));
+  }
+  const datetimeMatch = v.match(/^(\d{4})-(\d{2})-(\d{2})T.+$/);
+  if (datetimeMatch) {
+    // Datetime must have timezone offset (Z or ±HH:MM)
+    if (!/[Zz]$|[+-]\d{2}:\d{2}$/.test(v)) return false;
+    if (isNaN(Date.parse(v))) return false;
+    // Round-trip check on date portion to reject rolled-over datetimes
+    const [, yearStr, monthStr, dayStr] = datetimeMatch;
+    return isValidDateComponents(Number(yearStr), Number(monthStr), Number(dayStr));
+  }
+  return false;
+}
+
+function isValidDateComponents(year: number, month: number, day: number): boolean {
+  const d = new Date(year, month - 1, day);
+  return d.getFullYear() === year && d.getMonth() === month - 1 && d.getDate() === day;
+}
+
+export const strictDateSchema = z.string().refine(isStrictDateString, {
+  message: 'Date invalide (YYYY-MM-DD strict ou ISO datetime avec timezone)',
+});
